fix(accuracy): coerce numpy scalar scores to native float for serialization
The base Scorer.score() returned np.mean(...) (numpy.float64). After the
per-dataset-list refactor this raw score flows into Report.accuracy and
results.json, and both msgspec (result_summary.json) and json.dump
(results.json) reject numpy scalars — Report.to_json aborted the entire
report write with "Encoding objects of type numpy.float64 is unsupported"
(caught on a live gpt-oss-120b run). Fix at the source (return a native
float) and defensively at the entry boundary in _score_accuracy.

+++ src/inference_endpoint/evaluation/scoring.py
@@ -218,7 +218,10 @@ class Scorer(ABC):
             scores.append(self.score_single_sample(empirical[i], ground_truths[i]))
 
         n_repeats = len(scores) // self.dataset.num_samples()
-        return np.mean(scores), n_repeats
+        # float(...) so callers get a native Python float, not a numpy scalar:
+        # the score flows into result_summary.json (msgspec) and results.json
+        # (json), neither of which can serialize numpy.float64.
+        return float(np.mean(scores)), n_repeats
 
     def score_breakdown(self) -> dict[str, Any] | None:
         """Optional structured detail accompanying the scalar ``score()``.
